backend/indexer/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.17.0
requests==2.31.0
pydantic==2.5.0
